@@ -8,20 +8,33 @@ extension Internals.Override {
 
     #if DEBUG
     enum Raise {
-        typealias Closure = (Int32) -> Int32
+        typealias Closure = @Sendable (Int32) -> Int32
 
-        fileprivate static var closure: Closure = defaultClosure
+        fileprivate final class State: @unchecked Sendable {
+
+            var closure: Closure {
+                get { lock.withLock { _closure } }
+                set { lock.withLock { _closure = newValue } }
+            }
+
+            private let lock = Lock()
+            private var _closure: Closure = defaultClosure
+
+            init() {}
+        }
+
+        fileprivate static let state = State()
 
         private static let defaultClosure: Closure = {
             Foundation.raise($0)
         }
 
         static func replace(with closure: @escaping Closure) {
-            self.closure = closure
+            self.state.closure = closure
         }
 
         static func restore() {
-            closure = defaultClosure
+            state.closure = defaultClosure
         }
     }
     #endif
@@ -29,7 +42,7 @@ extension Internals.Override {
     @discardableResult
     static func raise(_ value: Int32) -> Int32 {
         #if DEBUG
-        Raise.closure(value)
+        Raise.state.closure(value)
         #else
         Foundation.raise(value)
         #endif
