@@ -13,18 +13,18 @@ struct DynamicValueMirror<Content: Sendable>: Sendable {
 
     // MARK: - Private properties
 
-    private let reflected: Mirror
+    private let reflected: @Sendable () -> Mirror
 
     // MARK: - Inits
 
     init(_ content: Content) {
-        reflected = Mirror(reflecting: content)
+        reflected = { Mirror(reflecting: content) }
     }
 
     // MARK: - Internal methods
 
     func callAsFunction() -> [Child] {
-        reflected.children.compactMap { child in
+        reflected().children.compactMap { child in
             (child.value as? DynamicValue).map {
                 .init(
                     label: child.label,
